fixed cbrt function

+++ src/GPU/CalculateEnergyCUDAKernel.cu
@@ -244,7 +244,7 @@ __global__ void BoxInterGPU(int *gpu_cellStartIndex,
   // 0 10 20 30 40
   // currentParticle 30
   //while(gpu_cellStartIndex[currentCell] < currentParticle) currentCell++;
-  int eCell = cbrt(numberOfCells);
+  int eCell = cbrt((float)numberOfCells);
   int x = (int)(gpu_x[currentParticle] / gpu_cell_x[0]);
   int y = (int)(gpu_y[currentParticle] / gpu_cell_y[1]);
   int z = (int)(gpu_z[currentParticle] / gpu_cell_z[2]);
